requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.9.2
python-dotenv==1.0.1

# Matching
sentence-transformers==3.0.1
scikit-learn==1.5.2
numpy==2.1.1
tabulate==0.9.0

# Optional Selenium fallback
selenium==4.24.0
webdriver-manager==4.0.2

# PDF/OCR and agents
pdfplumber==0.11.4
phidata==2.7.9
openai==1.51.2
pyngrok==7.2.3

# Firebase
firebase-admin==6.6.0

ddgs
playwright
